% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CATALYST_heatmap_plots.R
\name{colors_DAheatmap}
\alias{colors_DAheatmap}
\title{Plot the DA test heatmap.
Define colors for the experimental information of the DA heatmap}
\usage{
colors_DAheatmap(exp_info, comparison)
}
\arguments{
\item{exp_info}{Data.frame, with the experimental information besides the 
clusters IDs and the sample IDs}

\item{comparison}{Character, on which comparison the DA test has been performed.}
}
\value{
List of color object
}
\description{
Plot the DA test heatmap.
Define colors for the experimental information of the DA heatmap
}
\examples{
\dontrun{
# Extract values except cluster_id and sample_id
df <- as.data.frame(colData(sce)) \%>\% 
  dplyr::select(-cluster_id, -sample_id) \%>\% 
  unique()
# Order the values based on the comparison we want to test
df <- data.frame(df[order(df[,comparison]),], row.names = NULL)
# Generate the colors
colors_DAheatmap(exp_info = df, comparison)
}
}
