% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CyTOFtoolbox.R
\docType{package}
\name{CyTOFtoolbox}
\alias{CyTOFtoolbox}
\title{CyTOFtoolbox: A package regrouping tools to perform CyTOF data analysis.}
\description{
The CyTOFtoolbox package regroups functions to perfom:
clustering, differential expression analysis.
}
\section{Volcano plot functions}{

The volcano plot functions prepare the data to plot an adapted volcano plot.
It used the CytoGLMM output as well as the raw data.
}

