% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CATALYST_clusters_plots.R
\name{extract_SCE_info}
\alias{extract_SCE_info}
\title{Extract SCE information in preparation to plot the clusters.
Extract SCE information in preparation to plot the clusters.}
\usage{
extract_SCE_info(sce, metacluster, dr = "UMAP")
}
\arguments{
\item{sce}{SingleCellExperiment.}

\item{metacluster}{Character, metacluster to visualize.}

\item{dr}{Character, dimensionality reduction to use (by default dr = UMAP").}
}
\value{
Data.frame.
}
\description{
Extract SCE information in preparation to plot the clusters.
Extract SCE information in preparation to plot the clusters.
}
