% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_plot.R
\name{function_add_05}
\alias{function_add_05}
\title{Add 0.05 to the cyTOF values
(in prevention of the log transformation that will be apply)}
\usage{
function_add_05(x)
}
\arguments{
\item{x}{Numeric, vector}
}
\value{
Numeric, vector
}
\description{
Add 0.05 to the cyTOF values
(in prevention of the log transformation that will be apply)
}
