% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_plot.R
\name{function_combine_datas}
\alias{function_combine_datas}
\title{Combine data from CytoGLMM, the log2 fold change and the MSI}
\usage{
function_combine_datas(summary_CytoGLMM_fit, data_log2foldchange, data_MSI)
}
\arguments{
\item{summary_CytoGLMM_fit}{Tibble, formated results from the CytoGLMM model.}

\item{data_log2foldchange}{Tibble, log2 fold change per marker between 2 conditions.}

\item{data_MSI}{Tibble, Mean Signal Intensity (MSI) per marker.}
}
\value{
Tibble.
}
\description{
Combine data from CytoGLMM, the log2 fold change and the MSI
}
