% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_plot.R
\name{function_compute_log2foldchange}
\alias{function_compute_log2foldchange}
\title{Compute log2 fold change for each marker between 2 conditions}
\usage{
function_compute_log2foldchange(data, condition, protein_names)
}
\arguments{
\item{data}{Tibble, CyTOF data.}

\item{condition}{Character, column's name where the condition are stored.}

\item{protein_names}{Vector, vector of markers.}
}
\value{
Tibble.
}
\description{
Compute log2 fold change for each marker between 2 conditions
}
