% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_plot.R
\name{function_compute_MSI}
\alias{function_compute_MSI}
\title{Compute Mean Signal Intensity (MSI) per marker}
\usage{
function_compute_MSI(data, protein_names)
}
\arguments{
\item{data}{Tibble, CyTOF data.}

\item{protein_names}{Vector, vector of markers.}
}
\value{
Tibble.
}
\description{
Compute Mean Signal Intensity (MSI) per marker
}
