% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_plot.R
\name{function_prepare_output_CytoGLMMmodel}
\alias{function_prepare_output_CytoGLMMmodel}
\title{Prepare the output from the CytoGLMM model}
\usage{
function_prepare_output_CytoGLMMmodel(data, alpha)
}
\arguments{
\item{data}{Cytoglmm object, results from the CytoGLMM model.}

\item{alpha}{Numeric, threshold of p-value significance (by default alpha = 0.05)}
}
\value{
Tibble, formated CytoGLMM model results.
}
\description{
Prepare the output from the CytoGLMM model
}
