% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDS_plot.R
\name{plot_MDS_withthreshold}
\alias{plot_MDS_withthreshold}
\title{MDS on median marker expression}
\usage{
plot_MDS_withthreshold(
  df_samples,
  protein_names,
  sample_info_names,
  color,
  sample_label = "",
  cor_threshold = 0.5
)
}
\arguments{
\item{df_samples}{Data.frame, containing the cells intensities.}

\item{protein_names}{Vector, vector of markers.}

\item{sample_info_names}{Vector, vector of columns that are not marker's names.}

\item{color}{Character, indicates how to color the MDS dots.}

\item{sample_label}{Character, legend label.}

\item{cor_threshold}{Numeric, threshold on which marker to display in the circle
plot (by default cor_threshold = 0.5).}
}
\description{
MDS on median marker expression
}
