% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CATALYST_clusters_plots.R
\name{plot_metaclusters}
\alias{plot_metaclusters}
\title{Plot clusters.}
\usage{
plot_metaclusters(df, metacluster, point_size = 0.5, point_alpha = 0.5)
}
\arguments{
\item{df}{Data.frame, output of the extract_SCE_info function.}

\item{metacluster}{Character, metacluster to visualize.}

\item{point_size}{Numeric, (by default = 0.5)}

\item{point_alpha}{Numeric, (by default = 0.5)}
}
\value{
ggplot2 plot.
}
\description{
Plot clusters.
}
