% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CATALYST_heatmap_plots.R
\name{plotDAheatmap}
\alias{plotDAheatmap}
\title{Plot the DA test heatmap.
Plot the heatmap of the DA test.}
\usage{
plotDAheatmap(
  x,
  y,
  top_n = 20,
  comparison,
  all = FALSE,
  th = 0.05,
  normalize = TRUE,
  order = TRUE,
  show_sample_ID = TRUE
)
}
\arguments{
\item{x}{SingleCellExperiment.}

\item{y}{List, results of the diffcyt DA test.}

\item{top_n}{Numeric, number of top cluster to display (by default top_n = 20).}

\item{comparison}{Character, on which comparison the DA test has been performed.}

\item{all}{Logical, if TRUE all the clusters are displayed 
(if TRUE, top_n is ignored; by default all = FALSE).}

\item{th}{Numeric, p-value threshold (by default = 0.05).}

\item{normalize}{Logical, z-score normalization of the value. 
Relative population abundance are arcsine-square-root scaled prior to normalization.
(By default normalize = TRUE).}

\item{order}{Logical, order the results by significance (by default order = TRUE).}

\item{show_sample_ID}{Logical, specify is the sample IDs are displayed.
(By default show_sample_ID = TRUE).}
}
\value{
HeatmapList-class object.
}
\description{
Plot the DA test heatmap.
Plot the heatmap of the DA test.
}
