% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_plot.R
\name{prepare_data_for_volcanoplot}
\alias{prepare_data_for_volcanoplot}
\title{Prepare data for the volcano plot}
\usage{
prepare_data_for_volcanoplot(
  data,
  protein_names = NULL,
  condition,
  CytoGLMM_fit,
  alpha = 0.05
)
}
\arguments{
\item{data}{Tibble, raw CyTOF data.}

\item{protein_names}{Vector, vector of markers (by default = NULL, i.e. the 
protein names are extracted from the CytoGLMM object).}

\item{condition}{Character, column's name where the condition are stored.}

\item{CytoGLMM_fit}{Cytoglmm object, results from the CytoGLMM model.}

\item{alpha}{Numeric, threshold of p-value significance (by default alpha = 0.05)}
}
\value{
Tibble.
}
\description{
Prepare data for the volcano plot
}
