% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CyTOFtoolbox.R
\docType{package}
\name{CyTOFtoolbox}
\alias{CyTOFtoolbox}
\title{CyTOFtoolbox: A package regrouping visualization tools for CyTOF data analysis.}
\description{
The CyTOFtoolbox package regroups visualization tools for the analysis 
of CyTOF data.
}
\section{CytoGLMM-extension}{

The volcano plot functions prepare the data to plot a volcano plot adapted for
the CyTOF data. It uses the `CytoGLMM` output as well as the raw data.
A modified MDS plot function is also available, based on the original one from
the `CytoGLMM` package. It adds a threshold on the correlation marker/MDS axes.
}

