% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_plot.R
\name{volcano_plot}
\alias{volcano_plot}
\title{Volcano plot}
\usage{
volcano_plot(data, exp_conditions)
}
\arguments{
\item{data}{Tibble, output of the \code{prepare_data_for_volcanoplot} function.}

\item{exp_conditions}{Vector, conditions which are compared.}
}
\value{
ggplot2 object.
}
\description{
Volcano plot
}
